Added a function to the library that traverses a directory tree starting from a given point and executes certain commands in the following cases:
1. when a regular file is encountered
2. prior to entering and recursing inside a directory
3. after exiting a recursion inside a directory

The idea is that functionality can be built to perform user operations in
all files (and directories) in a directory tree. As an example, suppose
that every file needs to be hashed or encrypted inside the directory tree.
Another example is of having to systematically rename files and directories
in the directory tree. In both cases the user will have to author the
functions that are to be used in each case and then pass the pointers to
each as well as an appropriate argument for each to the function. Care must
be taken to make sure that the functions will not cause havoc and leave the
code in a place inside the filesystem that is problematic. Case must also
be taken such that the functions deal with null arguments (you have been
warned).
IN 2020/08/06 "master"